Removed pages outside of locale
