+++ app/[locale]/game/page.tsx
